refactor(ui): fix stale import, rename parse-date, extract timezone-combobox
- Fix StatusBar importing NLPDateInput from deleted @epicenter/ui/nlp-date-input
  path; consolidated to @epicenter/ui/natural-language-date-input
- Rename parse-date.ts → datetime-string.ts (file no longer parses dates,
  it serializes DateTimeString and gets local timezone)
- Move timezone-combobox.svelte to its own packages/ui/src/timezone-combobox/
  folder—it's a standalone reusable component, not NLP-specific
