+++ packages/ui/src/natural-language-date-input/index.ts
@@ -1,4 +1,3 @@
 export { default as NLPDateInput } from './nlp-date-input.svelte';
 export type { NLPDateInputProps } from './nlp-date-input.svelte';
-export { default as TimezoneCombobox } from './timezone-combobox.svelte';
-export { localTimezone, toDateTimeString } from './parse-date.js';
+export { localTimezone, toDateTimeString } from './datetime-string.js';
